Calculate grade range

This function calculates the best and worst possible grade averages that can be achieved based on the current grade average, the number of ECTS credits already completed, and the number of remaining ECTS credits to complete.

Usage

calculate_grade_range(
  completed_ects,
  current_grade,
  remaining_ects,
  max_passing_grade = 1,
  min_passing_grade = 5
)

Arguments

completed_ects

the number of ECTS credits already completed

current_grade

the current grade average

remaining_ects

the number of ECTS credits remaining to complete

max_passing_grade

the best possible grade

min_passing_grade

the worst possible grade

Value

a vector with the best and worst possible grade averages

Examples

calculate_grade_range(120, 3.5, 60)
#>  best_outcome worst_outcome 
#>      2.666667      4.000000 
calculate_grade_range(120, 3.5, 60, 1, 5)
#>  best_outcome worst_outcome 
#>      2.666667      4.000000